% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_base_filename.R
\name{filename_without_date}
\alias{filename_without_date}
\title{Filename without date}
\usage{
filename_without_date(filename)
}
\arguments{
\item{filename}{A string of the file name}
}
\value{
Filename with numbers, -, _ and common file extensions stripped out, and in lowercase
}
\description{
Filename without date
}
\examples{
filename_without_date("20220302_Some_File.csv")
#> [1] "somefile"

}
