% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_file.R
\name{get_keywords}
\alias{get_keywords}
\title{Get keywords}
\usage{
get_keywords(filename)
}
\arguments{
\item{filename}{String to get keywords from}
}
\value{
Character vector of keywords obtained from filename
}
\description{
Get keywords
}
\examples{
get_keywords("my_data apple friend-20220328.csv")

}
