% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_base_filename.R
\name{match_base_filename}
\alias{match_base_filename}
\title{Match base filename}
\usage{
match_base_filename(filename, files_to_match = NULL, recursive = FALSE)
}
\arguments{
\item{filename}{A string containing the full filepath you are looking for}

\item{files_to_match}{A character vector of files to match to filename. If NULL (default),
this is the list of files in the same directory as \code{filename}}

\item{recursive}{Boolean, default FALSE. Whether to search recursively for \code{files_to_match}
when \code{files_to_match} is not specified.}
}
\value{
A character vector of length one specifying the file in \code{files_to_match} which matches
filename.
}
\description{
Match base filename
}
\examples{
match_base_filename("20220302_Some_File.csv", c("20220304_somefile.csv", "20220302_Some_Other_File.csv"))

}
