% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_with_options.R
\name{read_custom_with_options}
\alias{read_custom_with_options}
\title{Read custom with options}
\usage{
read_custom_with_options(customloader, ...)
}
\arguments{
\item{customloader}{Function for reading a given filetype, e.g. readRDS}

\item{...}{Additional parameters to supply to customloader}
}
\value{
Function which reads that given filetype but first checks the file
exists and provides alternatives
}
\description{
Function factory which generates a function for reading in
a particular type of file.
}
\details{
The method for loading this file type is supplied as
customloader, along with any additional necessary parameters.

Useful if your file type is not Excel or csv but you still want
a generic function which loads the file but provides alternatives.
}
\examples{
read_rds_with_options <- read_custom_with_options(readRDS)

# myrds <- read_rds_with_options("/path/to/rds/myrds.rds")

}
