% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_file.R
\name{check_file}
\alias{check_file}
\title{Check File}
\usage{
check_file(filename, recursive = TRUE)
}
\arguments{
\item{filename}{A character vector of legth one}

\item{recursive}{Boolean. Whether to search for alternatives recursively or not}
}
\value{
A character vector
}
\description{
Check whether a file exists and if not suggest alternatives.
If the file exists the original filename is returned. If not
the user is prompted with alternative files based on the
keywords in the original filename, and they can select one
of these. The selected file is returned, or else an error
is thrown.
}
\examples{

# check_file("data/my_apples.csv")
}
