% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_all_excel_sheets.R
\name{read_all_excel_sheets}
\alias{read_all_excel_sheets}
\title{Read all Excel sheets}
\usage{
read_all_excel_sheets(xlsxFile, ...)
}
\arguments{
\item{xlsxFile}{String of path to Excel file}

\item{...}{Additional arguments to pass to openxlsx::read.xlsx}
}
\value{
List of contents of each Excel sheet
}
\description{
Reads all sheets of an Excel file to a list.
Checks the file exists and suggests alternatives if not.
}
\examples{

}
