% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_with_options.R
\name{read_csv_with_options}
\alias{read_csv_with_options}
\title{Read csv with options}
\usage{
read_csv_with_options(filename)
}
\arguments{
\item{filename}{A string of path to csv file}
}
\value{

}
\description{
Reads a csv file.
Checks first whether the file exists and suggests
alternatives if not.
}
\examples{

}
