% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_with_options.R
\name{read_excel_with_options}
\alias{read_excel_with_options}
\title{Read Excel with options}
\usage{
read_excel_with_options(filename, recursive = TRUE, ...)
}
\arguments{
\item{filename}{A string of path to Excel file}

\item{recursive}{Boolean. Whether to search for alternatives recursively or not}

\item{...}{Additional arguments to pass to readxl::read_excel}
}
\description{
Reads an Excel file using readxl::read_excel
Checks first whether the file exists and suggests
alternatives if not.
}
