% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_with_options.R
\name{read_excel_with_options}
\alias{read_excel_with_options}
\title{Read Excel with options}
\usage{
read_excel_with_options(filename)
}
\arguments{
\item{filename}{A string of path to Excel file}
}
\value{

}
\description{
Reads an Excel file.
Checks first whether the file exists and suggests
alternatives if not.
}
\examples{

}
