% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_alternative_files.R
\name{suggest_alternative_files}
\alias{suggest_alternative_files}
\title{Suggest alternative files}
\usage{
suggest_alternative_files(
  keywords,
  location,
  file_filter = "*",
  recursive = TRUE,
  useRstudio = TRUE
)
}
\arguments{
\item{keywords}{A character vector}

\item{location}{A character vector with one element}

\item{file_filter}{Glob filter to choose which type of files to look for e.g. "*.csv"}

\item{recursive}{Boolean for whether to recurse into directories to search}

\item{useRstudio}{Boolean for whether to use interactive R studio to find files}
}
\value{
A character vector or NULL
}
\description{
Searches for files in \code{location} containing one of the \code{keywords}, and returns
a character vector of the possibilties, prompting the user to pick one. Returns
the chosen possibility, or if none is chosen returns NA.
}
\examples{
# keywords <- c("apple", "orange")
# location <- "data/"
# suggest_alternative_files(keywords, location)
}
